services/ingestion_worker/requirements.txt
pika
pydantic
psycopg2-binary
elasticsearch
beautifulsoup4
markdownify
langchain-text-splitters
requests
